Ignore environment variables whose value is "none"
"none" is supposed to mean setting to the zero value for the relevant
config.  In confd all the Typha config fields are at their zero value
anyway, unless the environment configures them, so we can simply ignore
"none" values.

We _should_ ignore "none" values because otherwise we can try to look
for a Kubernetes service named "none".