For oneshot testing, run Typha each time after prepping datastore
Otherwise what happens is that Typha is already in-sync before the
calicoctl changes needed for the next test have gone through.  confd
thinks it can process templates as soon as the datastore is in sync, so
then we just get a race between confd and some of the resources getting
through.

With this change, the sequencing for each oneshot test is:

1. populate the datastore as needed for the test
2. start Typha
3. start confd
4. check templates
5. kill Typha

Then the first in-sync is guaranteed to include the datastore view that
the test needs.