Bug Fix: Stop using pointers for IPPools in IPAM
Fix a bug hit when filtering a slice and storing the result in a slice
of pointers.
The pointer that was being stored was the iteraction variable instead of
to the original data value.

The fix is to switch the code to using values instead of pointers as it
makes it harder to make this sort of mistake.