IPPool Create: Error if blocks conflict
When IPPools are deleted, blocks are left in place. Now that IPPools can
be created with different blockSizes, it's important that an
IPPool can't be created that would conflict with existing blocks that
use a different blockSize. If this was allowed, then overlapping blocks
would be created and IP address could be double assigned.

When creating an IPPool, check the list of existing blocks that are
covered by the pool being created. If they have a different block size
then throw an error.