Use unidirectional BGP connections in the node-to-node mesh.
Works around a bug in the BGP graceful restart protocol:
if a node is gracefully restarting, and its peer opens a connection
(due to its connection retry timer) before the restart has completed,
the graceful restart is aborted. This results in a route flap.

This change makes it so the BGP connection is only opened in one direction.
If a node is peering with another node that has a (lexicographically)
lesser IP address, we make the connection passive.

Tigera issue number: CNX-4867.