Bug Fix: List IPPools not defaulting blocksize properly
This affects upgrades from previous versions and was hit when trying to
allocate an IP (post upgrade) using CNI calico-ipam plugin.

I can't work out how to write an automated tests for this so I manually
tested it:

Against a blank etcd, I ran (using an old version pre-blocksize version
of calicoctl)
$ calicoctl apply -f ippool-noblock.yml
Successfully applied 1 'IPPool' resource(s)
$ ETCDCTL_API=3 etcdctl get /calico/resources/v3/projectcalico.org/ippools/my.ippool
/calico/resources/v3/projectcalico.org/ippools/my.ippool
{"kind":"IPPool","apiVersion":"projectcalico.org/v3","metadata":{"name":"my.ippool","uid":"575d6752-ce47-11e8-8a72-9cb6d089a923","creationTimestamp":"2018-10-12T17:50:43Z"},"spec":{"cidr":"192.168.0.0/24","ipipMode":"Never"}}

Then ran a CNI plugin without the fix, and then with the fix to verify
that it worked.
$ cat test.json | CNI_PATH=/tmp CNI_NETNS=/tmp CNI_IFNAME=eth0 CNI_COMMAND=ADD CNI_CONTAINERID=tom bin/amd64/calico-ipam
$ cat test.json
       │ File: test.json
   1   │ {
   2   │   "cniVersion": "0.2.0",
   3   │   "name": "net1",
   4   │   "type": "calico",
   5   │   "log_level": "DEBUG",
   6   │   "etcd_endpoints": "http://127.0.0.1:2379",
   7   │   "ipam": {
   8   │     "type": "calico-ipam",
   9   │     "subnet": "192.168.0.0/24"
  10   │   }
  11   │ }
  12   │

Without the fix I got:
{
    "code": 100,
    "msg": "failed to request 1 IPv4 addresses. IPAM allocated only 0"
}

With the fix I got:
{
    "cniVersion": "0.2.0",
    "ip4": {
        "ip": "192.168.0.192/32"
    },
    "dns": {}
}