Move the imageNames list from _config.yml to _data/imageNames.yml
For supporting the different image name for various versions, the
image name list was removed from global configuration _config.yml.
And add a configuration file imageNames.yml in _data folder for
supporting different versions.
By this way, we can unbind the imageName with versions. It can
support the different images in various versions.

Signed-off-by: Jingzhao <Jingzhao.Ni@arm.com>