Make dnsmasq invocation independent of Neutron DHCP code
In general, networking-calico's UT can break as changes are made to the
core Neutron code that some of our classes inherit from.  In particular,
the UT for _build_cmdline_callback recently broke because of this
change:

    0fce3ca2c1641fbcfb8327a86d7225e2c3972263
    Author: Jens Harbott <j.harbott@x-ion.de>  2018-10-29 17:08:33
    Committer: Brian Haley <bhaley@redhat.com>  2018-11-28 20:49:25

        Secure dnsmasq process against external abuse

        Currently any dhcp agent instance will work as an open resolver. For
        deployments using publicly routed addresses for tenant networks, this
        allows the agent being abused in dDoS attacks, see [1].

        By setting the `--local-service` option dnsmasq will filter DNS queries
        and reply only to queries from directly attached networks.

        [1] https://bugs.launchpad.net/neutron/+bug/1501206

        Closes-Bug: 1501206
        Change-Id: I76d810aad2ce0f15a88bd798963012fa0efca74e

We may want that change for Calico too; however, independently I'd like
to insulate networking-calico from being quietly broken by further
changes in this area, and we can do that by freezing in a copy of the
last-known-working Neutron code.

Change-Id: Ic9abdc641805455149c65b3f4502c1f4869dc0ec