Avoid ambiguous meaning for empty Region field
The new code here was consistently using it to mean "no region
configured", which maps to the string "no-region" in etcd key paths.

However, there is a convention that an empty XyzOptions field means that
the caller doesn't care about the value of that field, and that
conflicts with the "no region" meaning.

Resolve that here by changing the new XyzKey.Region and
XyzOptions.Region fields to XyzKey.RegionString and
XyzOptions.RegionString.  A region _string_ is always either "no-region"
or "region-<name>" and so is never validly empty.