Generate label with the OpenStack parent project ID for each VM
We previously generated VM endpoint labels indicating the project ID and
project name.  This change adds the parent project ID in the
'projectcalico.org/openstack-project-parent-id' label.

The parent project ID is never included in the Neutron request context,
so unfortunately we can now no longer use that context as a useful
source of project information.  If we haven't cached the information
already, we have to do a Keystone DB lookup.

Change-Id: Iaf7cb7c33c79a0592ea1722a280bc83ca0911fc5