fix a bug with alp manifests in helm
When helm functionality was added, it went unnoticed that some of the
ALP manifests rendered _only_ calico-node, and not the entire manifest
suite.

Fortunately, helm has a flag for the 'template' subcommand which only
renders a single manifest. This PR adds the ability to use that flag by
passing an arg in the '{% helm %}' tag