bpf: Add XDP program and means to build it
The program implements blocking the traffic by consulting the BPF maps
that serve as blacklists.

This also adds a docker file for building an image with clang, so we
can compile the C code into an ELF object file with the the BPF
bytecode.

The above are used in the Makefile.

Co-authored-by: Alban Crequy <alban@kinvolk.io>