git: Allow skipping some checks in pre-commit hook
The new `git-hooks/files-to-skip` file contains a list of files that
should not be checked for copyright or goimports. The obvious
candidates for skipping such checks are generated source files that
are not based on some template.

This commits adds one file to the list - the generated go code from
protobuf definitions.