config, dataplane: add parameters related to XDP functionality
XDPENABLED is about enabling XDP functionality (which is disabled by
default), provided that the system will meet certain requirements.

GENERICXDPENABLED is about allowing using the generic XDP mode, which is
the slowest one available. It is mostly used for debugging/testing.

XDPREFRESHINTERNAL is about how often the internal view of XDP state
in Felix should be synchronized with the actual state of XDP on the
system.

These get translated to internal dataplane options. The options are
not yet used, though.