dataplane, ipsets: add callbacks for some events in managers
These callbacks will be used by the XDP functionality coming in the
follow-up commit. The implementation is quite simple, since there will
be just one consumer of the callbacks functionality for now.

Later, if more consumers of callbacks appear, the implementation will
need to change to use slices to allow many callbacks for an
event. This kind of change should not change the API of callbacks,
though.

Also, for type-safety reasons, the code is quite repetitive, so maybe
writing a simple code-generator could be an option.