dataplane/xdp_test: Test XDP program with undesired mode
One test checks if resync catches the case when the installed XDP
program has an unallowed mode so resync marks the program for
replacement. Other test checks if the uninstall XDP program action
manages to remove the installed XDP with unallowed mode.

The unallowed mode here is the generic mode, so we change the way we
create the XDP IP state to allow only driver and offload modes.