dataplane/xdp: Handle XDP programs with unallowed mode
During resync routine, the logic should mark the XDP program for
replacement if it has unallowed mode.

During processing final actions, try all the modes, not only the
allowed ones, when trying to uninstall the program. That way we can
uninstall the XDP program marked for replacement by the resync logic.