bpf: Add sockops programs
The programs use the recent feature of the kernel used to accelerate
the local traffic.

There are two programs added: the decider and the redirector.

The decider implements logic for deciding which traffic can be
accelerated at a time when the connection is established. Currently
only traffic between the workload endpoint and the hardcoded address
being used by Envoy is allowed to be accelerated.

There is a need to differentiate the traffic coming from the app and
traffic coming from Envoy. Currently the heuristic is if source address
is the hardcoded address of Envoy, then the traffic comes from Envoy. If
neither addresses match the hardcoded address, then it comes from app.
The case when destination address is Envoy didn't happen to us in our
testing. The reason for such heuristics is that sockops program gets
executed before the packets are rewritten, so the packet from the app
still has the destination address of some other service, not of Envoy.

If traffic can be accelerated, a socket is put into the special
sockhash map.

The redirector does similar checks to sockops program to figure out
the key to use for looking up the socket in the sockhash map. If the
socket is found, then traffic can be accelerated.