dataplane: Hook sockmap into the internal dataplane
If SOCKMAPENABLED is true, the internal dataplane will try to create and
set up the sockmap state by plugging it into the endpoints manager with
callbacks.

Sockmap state processes the messages from endpoint manager about
workload endpoints to retrieve the IP address of the endpoint. The
address is then put into the endpoints map, so the sockops program
will know that the traffic related to that endpoint can be
accelerated. When a workload endpoint is removed, its address is also
dropped from the endpoints map.