dataplane/sockmap: address overlapping IPs
We can't rely on the updates we receive being ordered so we recalculate
the set of IPs every time we get a workload update, compare with the
actual state in the BPF map, and change accordingly.

This dumps the BPF state every time workloads change so it's not very
efficient but we don't expect frequent workload updates so it should be
fine.