Override etcd suboptions independent of the datastore type
There was previously an issue with the etcd suboptions not being applied from the felix configuration if the etcd datastore type wasn't set the etcd in the felix configuration as well. This is a problem because the default datastore type is etcd, so users are bothering to explicitly set the datastore type to etcd. Since the etcd suboptions only have an effect if the datastore is etcd we can just overide the suboptions with what's in the felix configuration, which fixes this issue.