Makefile logic to shell; vX.Y.Z checkout for networking-calico
When most of the logic in the Makefile is for PHONY targets, it's better
to move that to shell script; also it reduces the number of entrypoints
that people might think we have and support, and so helps future
maintainability.

Mixed in with that, I've added support for building a tagged release of
networking-calico, just by checking out that tag.