Fix for missing bin/calico-felix in wavetank run
The wavetank run failed in this way:

  + install -m 755 'bin/*' /tmp/rpmbuild/BUILDROOT/felix-3.9.0-0.1.0.dev.post46+20190807105811+0000+505f0a9.el7.x86_64/usr/bin/
  install: cannot stat 'bin/*': No such file or directory

And earlier it said:

  + git clean -ffxd -e bin/calico-felix
  Removing .go-pkg-cache/
  Removing bin/
  Removing vendor/

Whereas running on my own laptop I see:

  + git clean -ffxd -e bin/calico-felix
  Removing .go-pkg-cache/
  Removing bin/calico-felix-amd64
  Removing vendor/

git on wavetank is 2.11.0, whereas mine is 2.22.0, and the release notes
for 2.14.0 include:

 * "git clean -d" used to clean directories that has ignored files,
   even though the command should not lose ignored ones without "-x".
   "git status --ignored"  did not list ignored and untracked files
   without "-uall".  These have been corrected.

So I think the git clean on wavetank was ignoring the "-e
bin/calico-felix" and deleting that file.  Therefore let's not use git
clean, and instead delete the files and directories that need cleaning
explicitly.