Only remove interface routes that were previously added by Felix
The Felix agent is currently removing routes added by others to the
host side interface. This can be problematic for external
applications that want to also other things into pods/containers.

This change specifies a specific route protocol that will be added
to routes inserted by Felix. When removing routes it will only
remove routes that have that protocol.