bpf kube-proxy relacement internals
This is an initial commit with the implementation of the kube-proxy
replacement logic based on k8s 1.15 and able to write the state into bpf
maps.

The proxy.Proxy component can be used to implement a stand alone deamon
or to be used as part of another more complex one like felix.

Fixes BPF-106