Run CI against OpenStack stable/rocky (voting) and master (non-voting)
This change works around the current problem of networking-calico
being incompatible with OpenStack master CI (because of [1]), so that
we can get other useful changes into networking-calico.  It changes
the master devstack CI run to be non-voting, and adds a voting
devstack run against OpenStack Rocky.  I've chosen Rocky because

- it was the earliest release where the devstack code supports running
on Bionic, and the OpenStack CI VMs are all now Bionic

- networking-calico's tempest testing passes with Rocky.

But it's possible that we'd be OK with Stein as well.

(Train has just been released, and I'm pretty sure the latest breakage
[1] was introduced before that release, so I expect that we would not
be OK testing against Train.)

[1] I believe the problem in master is an incorrect assumption in
python-openstackclient, that the Neutron plugin provides the 'routes'
extension.  In fact that extension is optional and the Calico Neutron
plugin does not provide it.  I've posted
https://review.opendev.org/#/c/685312/ to correct that assumption.

Change-Id: I698f932d2d944053d04e6404d719c83734e41139