kube-proxy: dont leak derived svc on change
If a "derived" service changes because the original servie (clusterIP)
changed we should first delete the old one before creating a new one.

Since it isdifficult to spot this, whenever the original service
changes, we delete all derived ones. We always write the drived so this
will write the new ones.