bpf/proxy bug triggered by race if svc exist but no backend
It can happen that k8s tell us that there are no backends for a service
that has been removed but but we apply the change before k8s tell us
that the service was also removed. Once it tells us we do not remove the
backendsi due to the bug. Now we remove the backends correctly.

fixes BPF-147