Use Bionic PBR version for networking-calico instead of git describe
When there are multiple tags on the same commit, git describe does not
reliably use the most recent tag, as we'd like.  On the other hand, it
appears that the Bionic version of PBR does use the most recent tag,
so let's use that instead of git describe.  This has the additional
benefit that the Python egg version once installed is consistent with
the RPM and Debian package versions.

It also means that the networking-calico package version no longer
includes the date.  That's OK because it's not possible ever to delete
a commit from the networking-calico repo; and it's a minor advantage
because it means we don't re-upload to Launchpad when the code hasn't
actually changed.

Note: although Bionic PBR appears to do what we want here, PBR's
behaviour has changed over time:

Trusty PBR (0.7.0)
pbr_version=3.9.3.16.ga742d23

Xenial PBR (1.8.0)
pbr_version=3.10.2.dev13

Bionic PBR (3.1.1)
pbr_version=3.10.2.dev13

CentOS 7 PBR (0.8.0)
PBR_VERSION=3.10.2.dev13

Current latest PBR (5.4.3)
>>> import pbr.version
>>> pbr.version.VersionInfo('networking-calico').release_string()
'3.10.2.dev13'

We can't assume that any version of PBR would be OK, but it looks like
it has now given a stable result for some time.