VXLAN encap for NAT on host endpoint
When a packet should be NATed on a host endpoint, we NAT the packet
and we wrap the packet in vxlan to preserve the original source and
forward it to the dest pod where it gets decaped and consumed.

Response is encaped on the pod and sent back to the original host
which decaps the packet, does the SNAT and forwards it back to its
original source.

fixes BPF-88