bpf nodeports encap node-ip -> node-ip
When NATing to a nodeport from an external source, we encap with VXLAN
from the node-ip where it arrived to to node-ip where the dest pod
(service backend) is. The dest node decaps, passes it to the pod. On
reply, the pod encaps again, but for routing reasons, it uses the 1st
node's ip as dest and its own as source. The source is fixed when the
packet is leaving the node.

fixes BPF-159