pods can access nodeports
kube-proxy programs 255.255.255.255 as a special nodeport NAT entry. If
there is a NAT miss and the route is to a host, we search for this
special entry.

The downside is the special usecase for a broadcast address and the
fact that we cannot limit the nodeport IPs to certain CIDRs

fixes BPF-164