Do not NAT if externalTrafficPolicy is local
kubeproxy programs 0 backends if there are no local ones and thus we do
not NAT in such a case. We let the packets through if they pass policies
and the packets get dropped either by the host or the host responds RST
if there is no relevant service on the host.

endpoints of a services are sorted, local first, remote next.  This
allows the bpf code to pick either local or remote backend

calico_nat_v4_value->count and ->local both 32bit

32bit padding reserved