bpf NAT: affinity timeout as part of NAT frontend
If the affinityTimeo is non-zero, we create an entry with the selected
backend in the affinity table. If an entry already exists, and is not
too old yet, we use the backend from that entry instead. If it is
expired, we replace the backend with the newly picked backedend.

If there is no affinity timeout, we just return the selected backend.

fixes BPF-92