Expect same templates for both etcdv3 and kdd tests
The README.md indicates that this was always the intention, but it
appears that the templates were allowed to diverge at some point, with
different compiled_templates/etcdv3 and compiled_templates/kubernetes
directories.

I don't believe there's any strong reason for this, and the divergence
has allowed bugs where we add new templates for etcdv3 but forgot to
copy them for kubernetes as well.  It will be better for continued
development to force convergence again.