kube-proxy nodeport expansion fixer
When there is a miss, we remember the miss and we wait if some later rt
update would fix it. Whenever a new update(s) arrives, we do a pass over
the missed with the hope that we resolve it. If not, we sleep until
another update comes.

Since a new k8s state update of the services will trigger a whole new
scan of all services, we stop the fixer before applying the new state.