kube-proxy fix double read lock
With the WaitAfter ctx checker taking write lock to reliably unblock
the waiter, there is a race as if there is already a write waiting,
the next reader does not overtake it to avoid starvation. If we
already hold a reader lock, the writer will not get it and we cannot
get the second reader lock. WaitAfter provides a lockless lokkup
function to avoid this issue.