bpf TTL checks and decrement
TTL is checked to be > 1 and if not, we send ICMP TTL exceeded in
transit back where it came from.

If we bypass ipo stack by redirecting a packet straight to the next
iface, we must decrement the TTL by one ourselves. We know that it is
more than one by that time.

bpf VXLAN encap decrements TTL

When we put things in the tunnel, we must decrement TTL of the inner ip
header. The outer header inherits the TTL before decrement so that it
can be handled the same way as other traffic - either decremented by bpf
code or by the ip stack.

fixes BPF-89