bpf responds icmp TTL exceeded only just before NAT
If we do any sort of NATing, we know that the packet is being forwarded
and we check the TTL and respond ICMP TTL exceeded if need be as the
first thing before NATing as otherwise the reponse would either not pass
through or had post NAT ip header in it.

If the TTL gets to 1 and thus would be dropped on forward to the next
hop, we pass through the slow pass. If it it to the host would take this
path anyway.  If not, taking the slow path is ok as the performance does
not matter at this time so much. NATed triffic does not get this far,
see above.