bpf nodeport DSR
When a pod replies, we skip SNAT and encap where it would normally
happen with a return to the VXLAN tunnel and let the host endpoint do
the SNAT after all routing is done. The packet enters host namespace
with pod's ip to avoid failing rp_filter checks.

fixes BPF-253