Tell debhelper to use setup.py (and so not the Makefile)
Since we just added a Makefile for release purposes, we need to
prevent debhelpher from trying to use that when building packages -
which it will do by default in preference to setup.py.  This
incantation instructs it to use the Python build system,
i.e. setup.py, just as it was before we added that Makefile.