Generate networking-calico version similarly as for Felix
networking-calico was required, as an OpenStack project, to use a
thing called PBR ('Python Build Reasonableness') for the installed
networking-calico code to know its own version; in particular this
required the networking-calico version to follow a Semantic Versioning
format, which meant that it could not be the same format as the
versions we generate for Felix with the form

    ${last_tag}.post${commits_since}+${timestamp}+${sha}

for example `3.13.0-0.dev.post1+20200209224314+0000+c42f36b`.

Now that it's no longer an OpenStack project, I've removed the PBR
usage from inside networking-calico, and we can remove all the special
handling for the networking-calico version in this repo.  Then the
result will be networking-calico having versions with the same format
as Felix.

There's just one thing that we often need for networking-calico, but
can implement generally: ensuring that we use the right version when
there are multiple version tags on the same commit.  (In principle we
could one day need this for Felix too.)

Finally, in this commit, we standardize around having an initial 'v'
on version tags.  networking-calico tags from now on will have an
initial 'v', and the code is updated to expect that.  We also add a
'v' to the packaging changelog texts.