bpf: swap MAC addresses on redirect return
When we decide to immediatelly return a packet through the same iface,
we need to swap MACs. We do it only if we send ICMP error to WEP, but
PACKET_OTHEHOST is set because of the eth mismatch and ip_rcv_core
drops such packets as if the ifcase was in promisc mode.