bpf: fix GSO handling on the host side of NP
When a host iface receives a packet that is for nodeport, thus to be
encaped, it may be GSO due to the way it is received by the hardware and
we may see large packets due to receive side coallescing. In this case,
we can grow such packets to include the encap headers and forward them.

The test is the same as on the return from a workload.