Selector code refactor, step 1
Here we notice that RuleAPIToBackend has a lot of duplicated
selector-wrangling logic, for the Source and Destination cases, after
calling entityRuleAPIV2TOBackend, and that the net result
of (entityRuleAPIV2TOBackend + that logic) is just a single selector.

The extra logic is about how the namespace selector is combined in.

We move all of that inside a new func getSelector, that returns the
net result selector; then RuleAPIToBackend simply calls getSelector
twice.