bpf: conntrack table V2
Setup general code infrastructure to versioning maps. In principle,
the map name is extended by a single or a double digit number suffix.

In Golang, the the version can be bumped up by changing
MapParams.Version and a versioned map in C can be define using
CALI_MAP() macro that also defines version-less access functions to
the versioned map. This way no irrelevant code needs to be change
except bumping up the versions in C/bpf and golang.

Obviously, accessing new members etc. will be added/changed in the
current code, but access to the maps does not need to change.

This way upgraded felix does not use the old incompatible maps.

This does not implement any conversion from old to newer version.